feat: :zap: [BR-184] Comienzo de funcionalidad one page for checkout
