+++ app/core/components/checkoutPage/stripe.css
